requirements.txt
numpy
torch
transformers
open-clip-torch
opencv-python
Pillow
matplotlib
python-dotenv
yt-dlp
timm
safetensors
